Huge code moves to make things more structured.
